Moving LoadMore and Gallery components into utilities directory.
