+++ resources/assets/components/Gallery/PostGallery/PostGallery.js
@@ -3,8 +3,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import Card from '../../Card';
-import Gallery from '../Gallery';
-import LoadMore from '../../LoadMore';
+import Gallery from '../../utilities/Gallery/Gallery';
+import LoadMore from '../../utilities/LoadMore/LoadMore';
 import ReportbackItemContainer from '../../ReportbackItem';
 
 class PostGallery extends React.Component {
