+++ resources/assets/components/Gallery/SubmissionGallery/SubmissionGallery.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Gallery from '../Gallery';
 import { makeHash } from '../../../helpers';
+import Gallery from '../../utilities/Gallery/Gallery';
 import ReportbackItem from '../../ReportbackItem/ReportbackItem';
 
 class SubmissionGallery extends React.Component {
