+++ resources/assets/components/Gallery/index.js
@@ -1,5 +1,3 @@
 export * from './PostGallery';
 
 export * from './SubmissionGallery';
-
-export default from './Gallery';
