--- resources/assets/components/LoadMore/index.js
@@ -1,1 +1,0 @@
-export default from './LoadMore';
